feat: delay transmission for pdf generation (#224)
* Add retry policy for `WaitForPdfException` and implement PDF generation check

* Simplify condition for PDF transmission check in `StorageDialogportenDataMerger`

* Added retry policy and some cleanup

* Clean up

* Clean up

* Changed logic when waiting for PDF gen

* Fixed tests and logic error

* Null safety

* Add exception handling for `WaitForPdfException` to delay Transmission creation

* Add unit test and fix logic for counting PDFs per task in `AllPdfsGenerated`

* Move to DLQ for failed retires

* Merge main

* Fixed off by one. and make it end up in DLQ

* Link Pdf to task for when pdf was created

* refit workaround

* hm

* :)

* w

* Remove refit workaround

* Merge Main

* Clean up

* Clean up

+++ src/Altinn.DialogportenAdapter.WebApi/Features/Command/Sync/ISyncInstanceToDialogService.cs
@@ -14,7 +14,7 @@ namespace Altinn.DialogportenAdapter.WebApi.Features.Command.Sync;
 
 public interface ISyncInstanceToDialogService
 {
-    Task Sync(SyncInstanceCommand dto, CancellationToken cancellationToken = default);
+    Task Sync(SyncInstanceCommand dto, int currentAttempt = 1, CancellationToken cancellationToken = default);
 }
 
 internal sealed partial class SyncInstanceToDialogService : ISyncInstanceToDialogService
@@ -39,7 +39,7 @@ internal sealed partial class SyncInstanceToDialogService : ISyncInstanceToDialo
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
     }
 
-    public async Task Sync(SyncInstanceCommand dto, CancellationToken cancellationToken = default)
+    public async Task Sync(SyncInstanceCommand dto, int currentAttempt = 1, CancellationToken cancellationToken = default)
     {
         // To avoid performing requests needlessly, we attempt to load the app from cache first to see if
         // sync is disabled in which case we can skip the rest of the processing.
@@ -186,9 +186,17 @@ internal sealed partial class SyncInstanceToDialogService : ISyncInstanceToDialo
 
         // Create or update the dialog with the fetched data
         var mergeDto = new MergeDto(dialogId, existingDialog, application, applicationTexts, instance, events, dto.IsMigration || forceSilentUpsert);
-        var updatedDialog = await _dataMerger.Merge(mergeDto, cancellationToken);
+        var updatedDialog = await _dataMerger.Merge(mergeDto, currentAttempt, cancellationToken);
         var revision = await UpsertDialog(updatedDialog, existingDialog, syncAdapterSettings, dto.IsMigration || forceSilentUpsert, cancellationToken);
-
+        
+        // We throw an exception if PDFs are not generated and a form submission activity is present.
+        // Regardless of the current attempt so it will end up in DLQ
+        if (!StorageDialogportenDataMerger.AllPdfsGenerated(mergeDto) &&
+            updatedDialog.Activities.Any(activity => activity.Type is DialogActivityType.FormSubmitted))
+        {
+            throw new WaitForPdfException();
+        }
+        
         if (!syncAdapterSettings.DisableDelete && shouldDeleteAfterCreate && revision.HasValue)
         {
             var response = await _dialogportenApi.Delete(
